+++ docs/docs/coding/README.md
@@ -20,8 +20,8 @@ You will be able to encode and decode using the following:
 
 ```kotlin:ank
 import arrow.core.*
+import helios.*
 import helios.core.*
-import helios.annotations.*
 import helios.typeclasses.*
 //sampleStart
 val personJson = with(Person.encoder()) { Person("Abc", 10).encode() }
