+++ docs/docs/quickstart/README.md
@@ -10,8 +10,8 @@ Once Helios is added to your project, you can start adding the necessary imports
 
 ```kotlin:ank:silent
 import arrow.core.*
+import helios.*
 import helios.core.*
-import helios.annotations.*
 import helios.typeclasses.*
 ```
