+++ docs/src/main/kotlin/helios/sample/QuickStartHelper.kt
@@ -1,6 +1,6 @@
 package helios.core
 
-import helios.annotations.json
+import helios.json
 
 @json
 data class Person(val name: String, val age: Int) {
