+++ helios-optics/src/test/kotlin/helios/optics/domain.kt
@@ -1,6 +1,6 @@
 package helios.optics
 
-import helios.annotations.json
+import helios.json
 import helios.test.generators.alphaStr
 import io.kotlintest.properties.Gen
